Refactor: #113 navigation 디렉토리 ui 하위로 이동
