+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/carecall/screen/SetCallScreen.kt
@@ -41,7 +41,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
-import com.konkuk.medicarecall.navigation.Route
+import com.konkuk.medicarecall.ui.navigation.Route
 import com.konkuk.medicarecall.ui.common.component.CTAButton
 import com.konkuk.medicarecall.ui.feature.login.carecall.component.BenefitItem
 import com.konkuk.medicarecall.ui.feature.login.carecall.component.TimePickerBottomSheet
