+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/payment/screen/FinishSplashScreen.kt
@@ -21,7 +21,7 @@ import androidx.navigation.NavHostController
 import androidx.wear.compose.material3.Icon
 import androidx.wear.compose.material3.Text
 import com.konkuk.medicarecall.R
-import com.konkuk.medicarecall.navigation.navigateToMainAfterLogin
+import com.konkuk.medicarecall.ui.navigation.navigateToMainAfterLogin
 import com.konkuk.medicarecall.ui.common.component.CTAButton
 import com.konkuk.medicarecall.ui.type.CTAButtonType
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
