+++ app/src/main/java/com/konkuk/medicarecall/ui/feature/login/payment/screen/NaverPayScreen.kt
@@ -44,7 +44,7 @@ import androidx.core.net.toUri
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.konkuk.medicarecall.R
-import com.konkuk.medicarecall.navigation.Route
+import com.konkuk.medicarecall.ui.navigation.Route
 import com.konkuk.medicarecall.ui.feature.login.payment.viewmodel.NaverPayViewModel
 import com.konkuk.medicarecall.ui.model.PaymentResult
 import com.konkuk.medicarecall.ui.feature.settings.component.SettingsTopAppBar
