+++ app/src/main/java/com/konkuk/medicarecall/ui/navigation/BottomNavItem.kt
@@ -1,4 +1,4 @@
-package com.konkuk.medicarecall.navigation
+package com.konkuk.medicarecall.ui.navigation
 
 data class BottomNavItem(
     val label: String,
